Changes to be made
This folder is ardino IDE. since in my project i am using Node MCU ESP83266 that is why this code is important cause it will link to ThinkSpeak Cloud and will get Real time monitoring of Battery Voltage, Battery Percentage, etc. 
So the changes to be made in this code is you have to paste your API Code , Wi-Fi name and password.